Changed Style and ExStyle properties of window manager to use int type rather than long to match Win32 API specification.

+++ Rusty/Core/Common/WindowManager.cs
@@ -230,9 +230,9 @@ namespace IronAHK.Rusty
 
             public abstract bool Redraw();
 
-            public abstract long Style { get; set; }
+            public abstract int Style { get; set; }
 
-            public abstract long ExStyle { get; set; }
+            public abstract int ExStyle { get; set; }
 
             public abstract byte Transparency { get; set; }
